+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.8 (..., 2017)
 
+* Add Generics to IBaseRepository
 * Move BaseRepositoryFake to ptz-core-app
 
 ## 1.1.7 (March 30, 2017)